#188 - test refactor video image tags publishdate
