+++ tests/data/extractors/tags/test_tags_cnet.json
